pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scraper-hub"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.0",
    "requests>=2.31.0",
    "fake-useragent>=1.4.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["packages*"]
